Moving the traits files
It made no sense to put them where they were, so Ive moved them and
updated all the tests, On a close inspection they seem to be a
hinderance in traits more than anything, So I am going to be making them
As Classes and introducing a helper system to the application. This will
allow all helpers to be contained properly, used without `use`
statements and be highly extendable/configurable by the user with little
ease.